+++ src/main/java/com/pla/annoyingvillagers/procedures/BlueVillagerGeneralOnHurtProcedure.java
@@ -455,7 +455,7 @@ public class BlueVillagerGeneralOnHurtProcedure {
                     if (f <= 7.0F && levelaccessor instanceof Level) {
                         level = (Level)levelaccessor;
                         if (!level.isClientSide()) {
-                            level.playSound((Player)null, new BlockPos(d0, d1, d2), (SoundEvent)ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("minecraft", "entity.item.break")), SoundSource.NEUTRAL, 1.0F, 1.0F);
+                            level.playSound((Player)null, new BlockPos((int) d0, (int) d1, (int) d2), (SoundEvent)ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("minecraft", "entity.item.break")), SoundSource.NEUTRAL, 1.0F, 1.0F);
                         } else {
                             level.playLocalSound(d0, d1, d2, (SoundEvent)ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("minecraft", "entity.item.break")), SoundSource.NEUTRAL, 1.0F, 1.0F, false);
                         }
