+++ src/main/java/com/pla/annoyingvillagers/procedures/DarkObFarEntityOnHitEntityProcedure.java
@@ -8,6 +8,6 @@ import com.pla.annoyingvillagers.init.AnnoyingVillagersModBlocks;
 public class DarkObFarEntityOnHitEntityProcedure {
 
     public static void execute(LevelAccessor levelaccessor, double d0, double d1, double d2) {
-        levelaccessor.setBlock(new BlockPos(d0, d1, d2), ((Block) AnnoyingVillagersModBlocks.DARK_OB_UP.get()).defaultBlockState(), 3);
+        levelaccessor.setBlock(new BlockPos((int) d0, (int) d1, (int) d2), ((Block) AnnoyingVillagersModBlocks.DARK_OB_UP.get()).defaultBlockState(), 3);
     }
 }
