+++ src/main/java/com/pla/annoyingvillagers/procedures/GreenVillagerGeneralOnAttackingEntityProcedure.java
@@ -66,7 +66,7 @@ public class GreenVillagerGeneralOnAttackingEntityProcedure {
                     if (levelaccessor instanceof Level) {
                         level = (Level)levelaccessor;
                         if (!level.isClientSide()) {
-                            level.playSound((Player)null, new BlockPos(d0, d1, d2), (SoundEvent)ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("minecraft", "item.armor.equip_leather")), SoundSource.NEUTRAL, 1.0F, 1.0F);
+                            level.playSound((Player)null, new BlockPos((int) d0, (int) d1, (int) d2), (SoundEvent)ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("minecraft", "item.armor.equip_leather")), SoundSource.NEUTRAL, 1.0F, 1.0F);
                         } else {
                             level.playLocalSound(d0, d1, d2, (SoundEvent)ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("minecraft", "item.armor.equip_leather")), SoundSource.NEUTRAL, 1.0F, 1.0F, false);
                         }
@@ -140,7 +140,7 @@ public class GreenVillagerGeneralOnAttackingEntityProcedure {
                         if (levelaccessor instanceof Level) {
                             level = (Level)levelaccessor;
                             if (!level.isClientSide()) {
-                                level.playSound((Player)null, new BlockPos(d0, d1, d2), (SoundEvent)ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("minecraft", "item.armor.equip_leather")), SoundSource.NEUTRAL, 1.0F, 1.0F);
+                                level.playSound((Player)null, new BlockPos((int) d0, (int) d1, (int) d2), (SoundEvent)ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("minecraft", "item.armor.equip_leather")), SoundSource.NEUTRAL, 1.0F, 1.0F);
                             } else {
                                 level.playLocalSound(d0, d1, d2, (SoundEvent)ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("minecraft", "item.armor.equip_leather")), SoundSource.NEUTRAL, 1.0F, 1.0F, false);
                             }
