+++ src/main/java/com/pla/annoyingvillagers/procedures/GuardBreakProcedure.java
@@ -129,7 +129,7 @@ public class GuardBreakProcedure {
                 if (levelaccessor instanceof Level) {
                     level = (Level)levelaccessor;
                     if (!level.isClientSide()) {
-                        level.playSound((Player)null, new BlockPos(d0, d1, d2), (SoundEvent)ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation(AnnoyingVillagers.MODID, "s_g")), SoundSource.NEUTRAL, 2.0F, 1.0F);
+                        level.playSound((Player)null, new BlockPos((int) d0, (int) d1, (int) d2), (SoundEvent)ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation(AnnoyingVillagers.MODID, "s_g")), SoundSource.NEUTRAL, 2.0F, 1.0F);
                     } else {
                         level.playLocalSound(d0, d1, d2, (SoundEvent)ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation(AnnoyingVillagers.MODID, "s_g")), SoundSource.NEUTRAL, 2.0F, 1.0F, false);
                     }
@@ -138,7 +138,7 @@ public class GuardBreakProcedure {
                 if (levelaccessor instanceof Level) {
                     level = (Level)levelaccessor;
                     if (!level.isClientSide()) {
-                        level.playSound((Player)null, new BlockPos(d0, d1, d2), (SoundEvent)ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation(AnnoyingVillagers.MODID, "s_g_hit")), SoundSource.NEUTRAL, 3.0F, (float)Mth.nextDouble(AnnoyingVillagers.randomSource, 0.7D, 1.2D));
+                        level.playSound((Player)null, new BlockPos((int) d0, (int) d1, (int) d2), (SoundEvent)ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation(AnnoyingVillagers.MODID, "s_g_hit")), SoundSource.NEUTRAL, 3.0F, (float)Mth.nextDouble(AnnoyingVillagers.randomSource, 0.7D, 1.2D));
                     } else {
                         level.playLocalSound(d0, d1, d2, (SoundEvent)ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation(AnnoyingVillagers.MODID, "s_g_hit")), SoundSource.NEUTRAL, 3.0F, (float)Mth.nextDouble(AnnoyingVillagers.randomSource, 0.7D, 1.2D), false);
                     }
